matching-interim-api/requirements.txt
# requirements.txt - Production dependencies only

# Core FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# WSGI Server
gunicorn==21.2.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Data processing
pandas==2.2.0
numpy==1.26.3
openpyxl==3.1.2

# Configuration
pydantic==2.5.3
pydantic-settings==2.1.0

# Email validation
email-validator==2.1.0

# Utils
python-dateutil==2.8.2

# OpenAI (optional)
openai==1.12.0

# Monitoring
psutil==5.9.8